% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_case.R
\name{pathway_case}
\alias{pathway_case}
\title{Selecting pathway case with causal forests}
\usage{
pathway_case(
  data,
  treatment,
  outcome,
  num.trees = 2000,
  min.node.size = 1,
  honesty = TRUE,
  honesty.fraction = 0.5,
  seed = NULL,
  ci.group.size = 2,
  alpha = 0.05,
  imbalance.penalty = 0,
  clusters = NULL,
  sample.fraction = 0.5,
  mtry = floor(sqrt(ncol(data) - 2))
)
}
\arguments{
\item{data}{A data frame containing the variables for analysis. It must include
both the treatment and outcome variables, as well as any other covariates.}

\item{treatment}{The name of the treatment variable in the data frame.}

\item{outcome}{The name of the outcome variable in the data frame.}

\item{num.trees}{Number of trees to grow in the causal forest model. Defaults to 2000.}

\item{min.node.size}{Minimum size of the nodes in the causal forest. Defaults to 1.}

\item{honesty}{A logical value indicating whether to use honest estimation.
Defaults to TRUE.}

\item{honesty.fraction}{Fraction of the data to be used for estimation of
treatment effects in honest estimation. Defaults to 0.5.}

\item{seed}{An optional integer value for random seed to ensure reproducibility.
If NULL, the seed is not set. Defaults to NULL.}

\item{ci.group.size}{Size of the groups for confidence interval estimation.
Defaults to 2.}

\item{alpha}{Significance level for confidence interval calculation.
Defaults to 0.05.}

\item{imbalance.penalty}{A penalty term for imbalance in the splitting rule.
Defaults to 0.}

\item{clusters}{An optional vector of cluster IDs to be used for clustered sampling.
If NULL, no clustering is used. Defaults to NULL.}

\item{sample.fraction}{Fraction of the data to be used for building each tree.
Defaults to 0.5.}

\item{mtry}{Number of variables randomly sampled as candidates at each split.
Defaults to the floor of the square root of the number of variables in 'data' minus 2.}
}
\value{
A character string describing the observation number with the highest CATE,
the estimated CATE value for this observation, and its confidence interval.
}
\description{
This function performs a causal forest analysis using the 'grf' package to estimate
the Conditional Average Treatment Effects (CATE) on the provided data. It identifies
the observation with the highest estimated CATE, and returns its value along with a
confidence interval.
}
\examples{
# Assuming 'my_data' is a data frame with variables 'treatment_var' and 'outcome_var'
result <- pathway_case(my_data, "treatment_var", "outcome_var")
print(result)

}
