% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_outlier.R
\name{pathway_outlier}
\alias{pathway_outlier}
\title{Outlier-Adjusted Causal Forest Analysis}
\usage{
pathway_outlier(
  data,
  treatment,
  outcome,
  outlier_threshold,
  num.trees = 2000,
  min.node.size = 1,
  honesty = TRUE,
  honesty.fraction = 0.5,
  seed = NULL,
  ci.group.size = 2,
  alpha = 0.05,
  imbalance.penalty = 0,
  clusters = NULL,
  sample.fraction = 0.5,
  mtry = floor(sqrt(ncol(data) - 2))
)
}
\arguments{
\item{data}{A data frame containing the variables for analysis, including
treatment and outcome variables and other covariates.}

\item{treatment}{The name of the treatment variable in the data frame.}

\item{outcome}{The name of the outcome variable in the data frame.}

\item{outlier_threshold}{A numeric value representing the threshold for outlier
detection. Observations with treatment or outcome values beyond this threshold
will be considered outliers and excluded from the analysis.}

\item{num.trees}{Number of trees to grow in the causal forest model.
Defaults to 2000.}

\item{min.node.size}{Minimum size of the nodes in the causal forest.
Defaults to 1.}

\item{honesty}{A logical value indicating whether to use honest estimation.
Defaults to TRUE.}

\item{honesty.fraction}{Fraction of the data to be used for estimation of
treatment effects in honest estimation. Defaults to 0.5.}

\item{seed}{An optional integer value for random seed to ensure reproducibility.
If NULL, the seed is not set. Defaults to NULL.}

\item{ci.group.size}{Size of the groups for confidence interval estimation.
Defaults to 2.}

\item{alpha}{Significance level for confidence interval calculation.
Defaults to 0.05.}

\item{imbalance.penalty}{A penalty term for imbalance in the splitting rule.
Defaults to 0.}

\item{clusters}{An optional vector of cluster IDs to be used for clustered sampling.
If NULL, no clustering is used. Defaults to NULL.}

\item{sample.fraction}{Fraction of the data to be used for building each tree.
Defaults to 0.5.}

\item{mtry}{Number of variables randomly sampled as candidates at each split.
Defaults to the floor of the square root of the number of variables in 'data' minus 2.}
}
\value{
A character string describing the observation number with the highest CATE
among non-outliers, the estimated CATE value for this observation, and a statement
that this observation is among the non-outliers. The index of the observation is
with respect to the original dataset.
}
\description{
This function performs a causal forest analysis after adjusting for outliers in
the treatment and outcome variables. It identifies the observation with the
highest estimated Conditional Average Treatment Effect (CATE) among the non-outliers.
}
\details{
The function first excludes outliers in the treatment and outcome variables
based on the specified threshold. It then performs a causal forest analysis
using the 'grf' package on the remaining data to estimate CATEs.
}
