% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_time.R
\name{pathway_time}
\alias{pathway_time}
\title{Time Period Analysis of CATE with Causal Forest}
\usage{
pathway_time(data, id, time_period, treatment, outcome, num.trees = 2000)
}
\arguments{
\item{data}{A data frame containing the variables for analysis, including identifiers, time periods,
treatment and outcome variables, and other covariates.}

\item{id}{The name of the identifier variable in the data frame.}

\item{time_period}{The name of the time period variable in the data frame.}

\item{treatment}{The name of the treatment variable in the data frame.}

\item{outcome}{The name of the outcome variable in the data frame.}

\item{num.trees}{Number of trees to grow in the causal forest model. Defaults to 2000.}
}
\value{
A list containing two elements: 'plot', a ggplot object depicting the evolution of CATEs over
time for the ID with the highest overall CATE, and 'pred', a vector of the predicted CATE values for
this ID.
}
\description{
This function performs a causal forest analysis to estimate Conditional Average Treatment Effects
(CATE) over different time periods for each unique identifier in the dataset. It returns a plot
showing the evolution of CATEs over time for the ID with the highest overall CATE.
}
\details{
The function estimates CATEs separately for each unique identifier in the data, considering
the specified time period. It then identifies the identifier with the highest overall CATE and
generates a plot of CATEs across different time periods for this identifier.
}
\examples{
# Assuming 'my_data' is a data frame with variables 'ID', 'Time_Period', 'treatment_var', and 'outcome_var'
result <- pathway_time(my_data, "ID", "Time_Period", "treatment_var", "outcome_var")
print(result$plot)

}
