% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_top_cate.R
\name{plot_top_cate}
\alias{plot_top_cate}
\title{Plot Top N Cases Based on CATE}
\usage{
plot_top_cate(data, treatment, outcome, N = 10, num.trees = 2000, alpha = 0.05)
}
\arguments{
\item{data}{A data frame containing the variables for analysis, including
the treatment and outcome variables and other covariates.}

\item{treatment}{The name of the treatment variable in the data frame.}

\item{outcome}{The name of the outcome variable in the data frame.}

\item{N}{The number of top cases to display based on CATE. Defaults to 10.}

\item{num.trees}{Number of trees to grow in the causal forest model.
Defaults to 2000.}

\item{alpha}{Significance level for confidence interval calculation.
Defaults to 0.05.}
}
\value{
A ggplot object displaying the top N cases based on CATE, each represented by a point
with a corresponding confidence interval. Points are ordered by the magnitude of the CATE.
}
\description{
This function plots the top N cases based on Conditional Average Treatment Effects (CATE)
estimated using a causal forest model. It shows the CATE values along with their
95% confidence intervals for the selected top cases.
}
\details{
The function first fits a causal forest model to the data, then calculates the CATE
and its confidence intervals for each observation. It selects the top N cases based on
the highest CATE values and generates a plot displaying these values and their
confidence intervals.
}
\examples{
# Assuming 'my_data' is a data frame with variables 'treatment_var' and 'outcome_var'
plot_result <- plot_top_cate(my_data, "treatment_var", "outcome_var")
print(plot_result)

}
